requirements.txt
Click==7.0
Columnar==1.3.1
requests==2.22.0
python-dotenv==0.19.0
PyYAML==5.4.1
